handler restructuring
