+++ server/handler/catalog/get.go
@@ -1,0 +1,47 @@
+package catalog
+
+import (
+	db "SAU/mongod"
+	"SAU/server/model"
+	"context"
+	"net/http"
+	"time"
+
+	"github.com/gin-gonic/gin"
+	"github.com/sirupsen/logrus"
+)
+
+func GetAppByName(c *gin.Context, repository db.AppRepository) {
+	ctx, ctxErr := context.WithTimeout(c.Request.Context(), 30*time.Second)
+	defer ctxErr()
+
+	var appList []*model.App
+
+	//get parameter
+	appName := c.Query("app_name")
+
+	//request on repository
+	if result, err := repository.GetAppByName(appName, ctx); err != nil {
+		logrus.Error(err)
+	} else {
+		appList = result
+	}
+
+	c.JSON(http.StatusOK, gin.H{"apps": appList})
+}
+
+func GetAllApps(c *gin.Context, repository db.AppRepository) {
+	ctx, ctxErr := context.WithTimeout(c.Request.Context(), 30*time.Second)
+	defer ctxErr()
+
+	var appList []*model.App
+
+	//request on repository
+	if result, err := repository.Get(ctx); err != nil {
+		logrus.Error(err)
+	} else {
+		appList = result
+	}
+
+	c.JSON(http.StatusOK, gin.H{"apps": &appList})
+}
