+++ server/handler/info/info.go
@@ -1,0 +1,28 @@
+package info
+
+import (
+	"context"
+	"fmt"
+	"net/http"
+	"time"
+
+	"github.com/gin-gonic/gin"
+	"github.com/sirupsen/logrus"
+	"go.mongodb.org/mongo-driver/mongo"
+)
+
+func HealthCheck(c *gin.Context, client *mongo.Client) {
+	ctx, ctxCancel := context.WithTimeout(c.Request.Context(), 5*time.Second)
+	defer ctxCancel()
+
+	if ctx.Err() != nil {
+		err := fmt.Errorf("failed to create context with error: %v", ctx.Err())
+		logrus.Error(err)
+	}
+
+	if err := client.Ping(ctx, nil); err != nil {
+		c.JSON(http.StatusFailedDependency, gin.H{"status": "unhealthy"})
+	} else {
+		c.JSON(http.StatusOK, gin.H{"status": "healthy"})
+	}
+}
